Update patch set 1
Patch Set 1: (1 inline comment)

There are actually quite a few more strings from the English file that need to be translated. I've also been meaning to do a cleanup of the strings here, but I never got to it, and I can't go through everything at the moment. Will try to do so over the weekend.

Patch-set: 1